+++ packages/xrpl/src/models/methods/index.ts
@@ -1,5 +1,7 @@
 /* eslint-disable no-inline-comments -- Necessary for important note */
 /* eslint-disable max-lines -- There is a lot to export */
+import type { APIVersion, DEFAULT_API_VERSION } from '../common'
+
 import {
   AccountChannelsRequest,
   AccountChannelsResponse,
@@ -13,6 +15,8 @@ import {
   AccountInfoAccountFlags,
   AccountInfoRequest,
   AccountInfoResponse,
+  AccountInfoV1Response,
+  AccountInfoVersionResponseMap,
   AccountQueueData,
   AccountQueueTransaction,
 } from './accountInfo'
@@ -40,6 +44,8 @@ import {
 import {
   AccountTxRequest,
   AccountTxResponse,
+  AccountTxV1Response,
+  AccountTxVersionResponseMap,
   AccountTxTransaction,
 } from './accountTx'
 import { AMMInfoRequest, AMMInfoResponse } from './ammInfo'
@@ -76,11 +82,13 @@ import {
   LedgerQueueData,
   LedgerRequest,
   LedgerResponse,
+  LedgerV1Response,
   LedgerRequestExpandedTransactionsOnly,
   LedgerResponseExpanded,
   LedgerRequestExpandedAccountsAndTransactions,
   LedgerRequestExpandedAccountsOnly,
   LedgerRequestExpandedTransactionsBinary,
+  LedgerVersionResponseMap,
 } from './ledger'
 import { LedgerClosedRequest, LedgerClosedResponse } from './ledgerClosed'
 import { LedgerCurrentRequest, LedgerCurrentResponse } from './ledgerCurrent'
@@ -136,6 +144,8 @@ import { SubmitRequest, SubmitResponse } from './submit'
 import {
   SubmitMultisignedRequest,
   SubmitMultisignedResponse,
+  SubmitMultisignedV1Response,
+  SubmitMultisignedVersionResponseMap,
 } from './submitMultisigned'
 import {
   BooksSnapshot,
@@ -156,7 +166,7 @@ import {
   TransactionEntryRequest,
   TransactionEntryResponse,
 } from './transactionEntry'
-import { TxRequest, TxResponse } from './tx'
+import { TxRequest, TxResponse, TxV1Response, TxVersionResponseMap } from './tx'
 import {
   UnsubscribeBook,
   UnsubscribeRequest,
@@ -222,27 +232,27 @@ type Request =
 /**
  * @category Responses
  */
-type Response =
+type Response<Version extends APIVersion = typeof DEFAULT_API_VERSION> =
   // account methods
   | AccountChannelsResponse
   | AccountCurrenciesResponse
-  | AccountInfoResponse
+  | AccountInfoVersionResponseMap<Version>
   | AccountLinesResponse
   | AccountNFTsResponse
   | AccountObjectsResponse
   | AccountOffersResponse
-  | AccountTxResponse
+  | AccountTxVersionResponseMap<Version>
   | GatewayBalancesResponse
   | NoRippleCheckResponse
   // ledger methods
-  | LedgerResponse
+  | LedgerVersionResponseMap<Version>
   | LedgerClosedResponse
   | LedgerCurrentResponse
   | LedgerDataResponse
   | LedgerEntryResponse
   // transaction methods
   | SubmitResponse
-  | SubmitMultisignedResponse
+  | SubmitMultisignedVersionResponseMap<Version>
   | TransactionEntryResponse
   | TxResponse
   // path and order book methods
@@ -276,12 +286,15 @@ type Response =
   // Price Oracle methods
   | GetAggregatePriceResponse
 
-export type RequestResponseMap<T> = T extends AccountChannelsRequest
+export type RequestResponseMap<
+  T,
+  Version extends APIVersion = typeof DEFAULT_API_VERSION,
+> = T extends AccountChannelsRequest
   ? AccountChannelsResponse
   : T extends AccountCurrenciesRequest
   ? AccountCurrenciesResponse
   : T extends AccountInfoRequest
-  ? AccountInfoResponse
+  ? AccountInfoVersionResponseMap<Version>
   : T extends AccountLinesRequest
   ? AccountLinesResponse
   : T extends AccountNFTsRequest
@@ -291,7 +304,7 @@ export type RequestResponseMap<T> = T extends AccountChannelsRequest
   : T extends AccountOffersRequest
   ? AccountOffersResponse
   : T extends AccountTxRequest
-  ? AccountTxResponse
+  ? AccountTxVersionResponseMap<Version>
   : T extends AMMInfoRequest
   ? AMMInfoResponse
   : T extends GatewayBalancesRequest
@@ -357,15 +370,15 @@ export type RequestResponseMap<T> = T extends AccountChannelsRequest
   // then we'd get the wrong response type, LedgerResponse, instead of
   // LedgerResponseExpanded.
   T extends LedgerRequestExpandedTransactionsBinary
-  ? LedgerResponse
+  ? LedgerVersionResponseMap<Version>
   : T extends LedgerRequestExpandedAccountsAndTransactions
-  ? LedgerResponseExpanded
+  ? LedgerResponseExpanded<Version>
   : T extends LedgerRequestExpandedTransactionsOnly
-  ? LedgerResponseExpanded
+  ? LedgerResponseExpanded<Version>
   : T extends LedgerRequestExpandedAccountsOnly
-  ? LedgerResponseExpanded
+  ? LedgerResponseExpanded<Version>
   : T extends LedgerRequest
-  ? LedgerResponse
+  ? LedgerVersionResponseMap<Version>
   : T extends LedgerClosedRequest
   ? LedgerClosedResponse
   : T extends LedgerCurrentRequest
@@ -377,11 +390,11 @@ export type RequestResponseMap<T> = T extends AccountChannelsRequest
   : T extends SubmitRequest
   ? SubmitResponse
   : T extends SubmitMultisignedRequest
-  ? SubmitMultisignedResponse
+  ? SubmitMultisignedVersionResponseMap<Version>
   : T extends TransactionEntryRequest
   ? TransactionEntryResponse
   : T extends TxRequest
-  ? TxResponse
+  ? TxVersionResponseMap<Version>
   : T extends BookOffersRequest
   ? BookOffersResponse
   : T extends DepositAuthorizedRequest
@@ -420,20 +433,25 @@ export type RequestResponseMap<T> = T extends AccountChannelsRequest
   ? NFTsByIssuerResponse
   : T extends NFTHistoryRequest
   ? NFTHistoryResponse
-  : Response
+  : Response<Version>
 
 export type MarkerRequest = Request & {
   limit?: number
   marker?: unknown
 }
 
-export type MarkerResponse = Response & {
+export type MarkerResponse<
+  Version extends APIVersion = typeof DEFAULT_API_VERSION,
+> = Response<Version> & {
   result: {
     marker?: unknown
   }
 }
 
-export type RequestAllResponseMap<T> = T extends AccountChannelsRequest
+export type RequestAllResponseMap<
+  T,
+  Version extends APIVersion = typeof DEFAULT_API_VERSION,
+> = T extends AccountChannelsRequest
   ? AccountChannelsResponse
   : T extends AccountLinesRequest
   ? AccountLinesResponse
@@ -442,14 +460,12 @@ export type RequestAllResponseMap<T> = T extends AccountChannelsRequest
   : T extends AccountOffersRequest
   ? AccountOffersResponse
   : T extends AccountTxRequest
-  ? AccountTxResponse
+  ? AccountTxVersionResponseMap<Version>
   : T extends LedgerDataRequest
   ? LedgerDataResponse
-  : T extends AccountTxRequest
-  ? AccountTxResponse
   : T extends BookOffersRequest
   ? BookOffersResponse
-  : MarkerResponse
+  : MarkerResponse<Version>
 
 export {
   // Allow users to define their own requests and responses.  This is useful for releasing experimental versions
@@ -467,6 +483,7 @@ export {
   AccountInfoAccountFlags,
   AccountInfoRequest,
   AccountInfoResponse,
+  AccountInfoV1Response,
   AccountQueueData,
   AccountQueueTransaction,
   AccountLinesRequest,
@@ -484,6 +501,7 @@ export {
   AccountOffersResponse,
   AccountTxRequest,
   AccountTxResponse,
+  AccountTxV1Response,
   AccountTxTransaction,
   GatewayBalance,
   GatewayBalancesRequest,
@@ -495,6 +513,7 @@ export {
   // ledger methods
   LedgerRequest,
   LedgerResponse,
+  LedgerV1Response,
   LedgerQueueData,
   LedgerBinary,
   LedgerModifiedOfferCreateTransaction,
@@ -514,10 +533,12 @@ export {
   SubmitResponse,
   SubmitMultisignedRequest,
   SubmitMultisignedResponse,
+  SubmitMultisignedV1Response,
   TransactionEntryRequest,
   TransactionEntryResponse,
   TxRequest,
   TxResponse,
+  TxV1Response,
   // path and order book methods with types
   BookOffersRequest,
   BookOffer,